# docs/openapi/public.yaml
openapi: 3.1.0
info:
  title: City Flood MVP API
  version: 0.1.0
paths:
...
  /v1/cities:
    get:
      tags:
      - cities
      summary: Cities
      operationId: cities_v1_cities_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CityOut'
                type: array
                title: Response Cities V1 Cities Get
              example:
              - city_id: city_porto_mvp
                name: MVP City
                country: PT
        '400':
          $ref: '#/components/responses/InvalidArgument'
  /v1/cities/{city_id}:
    get:
      tags:
      - cities
      summary: City
      operationId: city_v1_cities__city_id__get
      parameters:
      - name: city_id
        in: path
        required: true
        schema:
          type: string
          title: City Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityOut'
              example:
                city_id: city_porto_mvp
                name: MVP City
                country: PT
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/cities/{city_id}/assets:
    get:
      tags:
      - assets
      summary: City Assets
      operationId: city_assets_v1_cities__city_id__assets_get
      parameters:
      - name: city_id
        in: path
        required: true
        schema:
          type: string
          title: City Id
      - name: type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: asset_type filter
          title: Type
        description: asset_type filter
      - name: bbox
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: minLon,minLat,maxLon,maxLat
          title: Bbox
        description: minLon,minLat,maxLon,maxLat
      - name: near
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: lon,lat for radial search
          title: Near
        description: lon,lat for radial search
      - name: radius_m
        in: query
        required: false
        schema:
          type: number
          maximum: 50000
          minimum: 1
          default: 500.0
          title: Radius M
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: tag filter key:value (repeatable)
          title: Tag
        description: tag filter key:value (repeatable)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 5000
          minimum: 1
          default: 2000
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetOut'
                title: Response City Assets V1 Cities  City Id  Assets Get
              example:
              - asset_id: pipe_123
                city_id: city_porto_mvp
                asset_type: pipe
                name: Pipe 0001
                geom_geojson:
                  type: LineString
                  coordinates:
                  - - -8.69
                    - 41.07
                  - - -8.68
                    - 41.08
                props:
                  diameter_m: 0.5
                  material: PVC
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/assets/{asset_id}:
    get:
      tags:
      - assets
      summary: Asset
      operationId: asset_v1_assets__asset_id__get
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetOut'
              example:
                asset_id: pipe_123
                city_id: city_porto_mvp
                asset_type: pipe
                name: Pipe 0001
                geom_geojson:
                  type: LineString
                  coordinates:
                  - - -8.69
                    - 41.07
                  - - -8.68
                    - 41.08
                props:
                  diameter_m: 0.5
                  material: PVC
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/assets/{asset_id}/upstream:
    get:
      tags:
      - assets
      summary: Asset Upstream
      operationId: asset_upstream_v1_assets__asset_id__upstream_get
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      - name: depth
        in: query
        required: false
        schema:
          type: integer
          maximum: 20
          minimum: 1
          default: 3
          title: Depth
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetOut'
                title: Response Asset Upstream V1 Assets  Asset Id  Upstream Get
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/assets/{asset_id}/downstream:
    get:
      tags:
      - assets
      summary: Asset Downstream
      operationId: asset_downstream_v1_assets__asset_id__downstream_get
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      - name: depth
        in: query
        required: false
        schema:
          type: integer
          maximum: 20
          minimum: 1
          default: 3
          title: Depth
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetOut'
                title: Response Asset Downstream V1 Assets  Asset Id  Downstream Get
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /v1/telemetry/events:batch:
    post:
      tags:
      - ingest
      summary: Ingest Events Batch
      operationId: ingest_events_batch_v1_telemetry_events_batch_post
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedBearer'
        '403':
          $ref: '#/components/responses/ForbiddenBearer'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
      security:
      - BearerAuth: []
...
components:
  schemas:
...
    AssetOut:
      properties:
        asset_id:
          type: string
          title: Asset Id
        city_id:
          type: string
          title: City Id
        asset_type:
          type: string
          enum:
          - river_segment
          - node
          - pipe
          - outfall
          - rain_gauge
          - river_gauge
          - tank
          - pump
          title: Asset Type
        name:
          type: string
          title: Name
        geom_geojson:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Geom Geojson
        props:
          additionalProperties: true
          type: object
          title: Props
      type: object
      required:
      - asset_id
      - city_id
      - asset_type
      - name
      title: AssetOut
    CityOut:
      properties:
        city_id:
          type: string
          title: City Id
        name:
          type: string
          title: Name
        country:
          type: string
          title: Country
      type: object
      required:
      - city_id
      - name
      - country
      title: CityOut
...
    IngestEventIn:
      properties:
        ts:
          type: string
          format: date-time
          title: Ts
        type:
          type: string
          title: Type
        value:
          type: number
          title: Value
        quality_flag:
          type: string
          title: Quality Flag
          default: ok
      type: object
      required:
      - ts
      - type
      - value
      title: IngestEventIn
    IngestEventResult:
      properties:
        ts:
          type: string
          format: date-time
          title: Ts
        type:
          type: string
          title: Type
        status:
          type: string
          enum:
          - accepted
          - rejected
          title: Status
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      type: object
      required:
      - ts
      - type
      - status
      title: IngestEventResult
    IngestRequest:
      properties:
        device_id:
          type: string
          title: Device Id
        sent_at:
          type: string
          format: date-time
          title: Sent At
        events:
          items:
            $ref: '#/components/schemas/IngestEventIn'
          type: array
          title: Events
      type: object
      required:
      - device_id
      - sent_at
      - events
      title: IngestRequest
    IngestResponse:
      properties:
        device_id:
          type: string
          title: Device Id
        received:
          type: integer
          title: Received
        accepted:
          type: integer
          title: Accepted
        rejected:
          type: integer
          title: Rejected
        results:
          items:
            $ref: '#/components/schemas/IngestEventResult'
          type: array
          title: Results
      type: object
      required:
      - device_id
      - received
      - accepted
      - rejected
      - results
      title: IngestResponse
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            request_id:
              type: string
          required:
          - code
          - message
          - details
          - request_id
      required:
      - error
  responses:
    InvalidArgument:
      description: 'INVALID_ARGUMENT: Validation error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INVALID_ARGUMENT
              message: Validation error
              details: {}
              request_id: req_example
...
    UnauthorizedBearer:
      description: 'UNAUTHORIZED: Unauthorized'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Unauthorized
              details: {}
              request_id: req_example
    ForbiddenBearer:
      description: 'FORBIDDEN: Forbidden'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: Forbidden
              details: {}
              request_id: req_example
    NotFound:
      description: 'NOT_FOUND: Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
              details: {}
              request_id: req_example
...
    UnprocessableEntity:
      description: 'UNPROCESSABLE_ENTITY: Unprocessable entity'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNPROCESSABLE_ENTITY
              message: Unprocessable entity
              details: {}
              request_id: req_example
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security: []